requirements.txt
fastapi
uvicorn
python-dotenv
sqlite3